DispatchQueue: Move from msg/simple to msg
(cherry picked from commit f282ad053404dee538da84f03b9f9378ef86b26e)
