+++ src/CMakeLists.txt
@@ -263,7 +263,7 @@ set(libcommon_files
   mon/MonClient.cc
   mon/MonMap.cc
   msg/simple/Accepter.cc
-  msg/simple/DispatchQueue.cc
+  msg/DispatchQueue.cc
   msg/Message.cc
   osd/ECMsgTypes.cc
   osd/HitSet.cc
